requirements.txt
uvicorn=0.34.0
wget=3.2
tabula-py=2.10.0
fastapi=0.115.12
pandas=2.2.3